Add lines module and differentiate Path/Line
